@@ -1,11 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 34
+    namespace 'com.cms.canteen.foodmanagementapp'
     defaultConfig {
-        applicationId "com.example.hp.myapplication"
+        applicationId "com.cms.canteen.foodmanagementapp"
         minSdkVersion 15
-        targetSdkVersion 27
+        multiDexEnabled true
+        targetSdkVersion 34
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -22,26 +24,22 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.cepheuen.elegant-number-button:lib:1.0.2'
 
-    //implementation 'com.cepheuen.elegant-number-button:lib:1.0.2'
-    //implementation 'com.android.support:design:26.+'
-
     implementation 'com.amulyakhare:com.amulyakhare.textdrawable:1.0.1'
     implementation 'com.readystatesoftware.sqliteasset:sqliteassethelper:+'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
     implementation 'com.android.support.constraint:constraint-layout:1.1.0'
     implementation 'com.squareup.picasso:picasso:2.71828'
-    implementation 'com.android.support:design:27.1.1'
+    implementation 'com.google.android.material:material:1.11.0'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.1'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
 
     implementation 'com.google.firebase:firebase-core:12.0.1'
     implementation 'com.google.firebase:firebase-database:12.0.1'
-    //compile 'info.hoang8f:fbutton:1.0.5'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.viewpager:viewpager:1.0.0'
     implementation 'com.rengwuxian.materialedittext:library:2.1.4'
 
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
     implementation 'com.firebaseui:firebase-ui-database:1.2.0'
     implementation 'com.github.bumptech.glide:glide:3.7.0'
 }
